Начало.

+++ named-server/src/main/java/ru/pastor/templates/named/configuration/CacheConfiguration.java
@@ -19,22 +19,22 @@ import java.time.Duration;
 public class CacheConfiguration {
 
   @Bean("NamedCache.Redis")
-  public NamedCache<String, Long> redisNamedCache(ReactiveRedisOperations<String, Long> operations) {
+  public NamedCache<String, Integer> redisNamedCache(ReactiveRedisOperations<String, Integer> operations) {
     return new Redis(operations);
   }
 
   @Bean("NamedCache.Values")
-  public NamedCache<String, Long> valuesNamedCache(MeterRegistry registry,
-                                                   @Qualifier("NamedCache.Redis") NamedCache<String, Long> redisCache) {
+  public NamedCache<String, Integer> valuesNamedCache(MeterRegistry registry,
+                                                   @Qualifier("NamedCache.Redis") NamedCache<String, Integer> redisCache) {
     return new NamedCache.Local<>("values", Duration.ofHours(1), registry, redisCache, null);
   }
 
   @Bean("NamedCache.Catalogue")
-  public NamedCache<String, Long> catalogueNamedCache(MeterRegistry registry,
+  public NamedCache<String, Integer> catalogueNamedCache(MeterRegistry registry,
                                                       CatalogueRepository repository) {
     return new NamedCache.Local<>("catalogue", Duration.ofHours(4), registry, new NamedCache.ReadOnly<>() {
       @Override
-      public Mono<Long> get(String key) {
+      public Mono<Integer> get(String key) {
         return repository.get(key).map(CatalogueEntity::id);
       }
     }, entry ->
